Add 13-point value scoring system with API endpoint and frontend page
Implements a company scoring system based on tsx-aggregator's value investing
criteria. Includes data models, SQL data access with fallback concept chains
(33 US-GAAP concepts), scoring computation service with pure-function
architecture, REST endpoint (GET /api/companies/{cik}/scoring), and Angular
scoring page with scorecard table, derived metrics, and raw data view.

Also fixes GetCompanyByIdStmt NpgsqlParameter<ulong> bug (cast to long),
and changes scoring query to use DISTINCT ON with max end_date instead of
dp.end_date = report_date to correctly handle XBRL comparative period data.

+++ dotnet/Stocks.Persistence/Database/Statements/GetCompanyByIdStmt.cs
@@ -38,7 +38,7 @@ WHERE company_id = @company_id;
     protected override void ClearResults() => Company = Company.Empty;
 
     protected override IReadOnlyCollection<NpgsqlParameter> GetBoundParameters() =>
-        [new NpgsqlParameter<ulong>("company_id", _companyId)];
+        [new NpgsqlParameter<long>("company_id", (long)_companyId)];
 
     protected override bool ProcessCurrentRow(NpgsqlDataReader reader) {
         Company = new Company(
